Code improvments. Some user-important changes were made
CHANGES:
Old variables are deleted now.
Everything(probably) is based on selection.

DESCRIPTION:
Caret is mostly working as expected.
One exception is when we place cursor inside full word(that doesn't
require correction) then after ctrl+space combination next word is
automaticlly sugested, and we don't want that.
We want to override this word! But that's very tiny problem, easy to
solve.